requirements.txt
torch>=2.7.1
torchvision>=0.22.1
numpy>=2.0.2
scipy>=1.13.1
matplotlib>=3.9.4
prettytable>=3.12.0
ujson>=5.10.0
pyyaml>=6.0.2
pynvml>=11.5.0
tqdm>=4.67.1
requests>=2.32.3
pillow>=11.3.0
six>=1.17.0
python-dateutil>=2.9.0.post0
pytz>=2024.2
scikit-learn>=1.5.2
pandas>=2.2.3
